linux-gen: dpdk: cast addresses to uintptr_t rather than uint64_t
This is to remove the following error:

pktio/dpdk.c: In function ‘mbuf_data_off’:
pktio/dpdk.c:126:9: error: cast from pointer to integer of different
size [-Werror=pointer-to-int-cast]
  return (uint64_t)pkt_hdr->buf_hdr.seg[0].data -
         ^
pktio/dpdk.c:127:4: error: cast from pointer to integer of different
size [-Werror=pointer-to-int-cast]
    (uint64_t)mbuf->buf_addr;

Reviewed-and-tested-by: Matias Elo <matias.elo@nokia.com>

+++ platform/linux-generic/pktio/dpdk.c
@@ -123,8 +123,8 @@ static unsigned cache_size(uint32_t num)
 static inline uint16_t mbuf_data_off(struct rte_mbuf *mbuf,
 				     odp_packet_hdr_t *pkt_hdr)
 {
-	return (uint64_t)pkt_hdr->buf_hdr.seg[0].data -
-			(uint64_t)mbuf->buf_addr;
+	return (uintptr_t)pkt_hdr->buf_hdr.seg[0].data -
+			(uintptr_t)mbuf->buf_addr;
 }
 
 /**
